feat: Implementing cross-segment read/write for WAL based on local disk (#1556)
## Rationale
Improving WAL based on local disk.

This is a follow-up task for #1552.

## Detailed Changes
1. Make MAX_FILE_SIZE configurable.
2. Allocate enough space when creating a segment to avoid remapping when
appending to the segment.​
3. Add `MultiSegmentLogIterator` to enable cross-segment reading.
4. When writing, if the current segment has insufficient space, create a
new segment and write to the new segment.​

## Test Plan
Unit test.

+++ src/wal/src/local_storage_impl/config.rs
@@ -20,7 +20,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct LocalStorageConfig {
     pub path: String,
-    pub max_segment_size: u64,
+    pub max_segment_size: usize,
     pub cache_size: usize,
 }
